reorganize package + fix installation
