+++ CMakeLists.txt
@@ -17,7 +17,17 @@ ADD_REQUIRED_DEPENDENCY("qpOASES")
 
 INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/include)
 
+SET(${PROJECT_NAME}_HEADERS
+  include/ddp-actuator-solver/costfunction.hh
+  include/ddp-actuator-solver/dynamicmodel.hh
+  include/ddp-actuator-solver/ddpsolver.hh)
+
+SET(HEADERS
+  ${${PROJECT_NAME}_HEADERS}
+  )
+
 ADD_SUBDIRECTORY(src)
 ADD_SUBDIRECTORY(test)
+ADD_SUBDIRECTORY(examples)
 
 SETUP_PROJECT_FINALIZE()
