+++ examples/romeo_actuator/CMakeLists.txt
@@ -1,0 +1,4 @@
+_local_test_ddp(  main
+		              mainMPC
+		              testModel)
+#ADD_EXECUTABLE(benchControl benchControl.cpp ${PROJECT_NAME})
