+++ examples/romeo_actuator/romeosimpleactuator.hh
@@ -1,12 +1,12 @@
-#ifndef DCTEMP_H
-#define DCTEMP_H
+#ifndef ROMEOSIMPLEACTUATOR_H
+#define ROMEOSIMPLEACTUATOR_H
 
-#include "dynamicmodel.h"
+#include <ddp-actuator-solver/dynamicmodel.hh>
 
-class DCTemp : public DynamicModel<double,5,1>
+class RomeoSimpleActuator : public DynamicModel<double,4,1>
 {
 public:
-    DCTemp(double& mydt,bool noiseOnParameters=0);
+    RomeoSimpleActuator(double& mydt,bool noiseOnParameters=0);
 private:
 protected:
 
@@ -15,15 +15,27 @@ public:
 private:
     double dt;
 private:
-    double J;
-    double K_M;
-    double f_VL;
-    double R_th;
-    double tau_th;
+    double k;
+    double R;
+    double Jm;
+    double Jl;
+    double fvm;
+    double fvl;
+    double Kt;
+    double mu;
+    double Cf0;
+    double a;
 private:
-    stateVec_t Xreal,dX;
-    stateVec_t x_next,k1,k2,k3,k4;
+    stateVec_t Xreal;
     stateMat_t Id;
+    stateMat_t A;
+    stateMat_t Ad;
+    stateR_commandC_t B;
+    stateR_commandC_t Bd;
+    double A13atan;
+    double A33atan;
+    stateMat_t fxBase;
+    stateR_commandC_t fuBase;
 
     stateMat_t QxxCont;
     commandMat_t QuuCont;
@@ -32,7 +44,6 @@ private:
 protected:
     // methods //
 public:
-    stateVec_t computeDeriv(double& dt, const stateVec_t& X, const commandVec_t &U);
     stateVec_t computeNextState(double& dt, const stateVec_t& X, const commandVec_t &U);
     void computeAllModelDeriv(double& dt, const stateVec_t& X, const commandVec_t &U);
     stateMat_t computeTensorContxx(const stateVec_t& nextVx);
