+++ examples/temperature_control/CMakeLists.txt
@@ -1,0 +1,2 @@
+
+_local_test_ddp(mainTemp)
