--- test/BenchCom.cpp
@@ -1,5 +1,0 @@
-//
-// Created by flo on 11/01/17.
-//
-
-#include "BenchCom.h"
