--- test/BenchCom.h
@@ -1,16 +1,0 @@
-//
-// Created by flo on 11/01/17.
-//
-
-#ifndef PROJECT_BENCHCOM_H
-#define PROJECT_BENCHCOM_H
-
-
-
-class BenchCom {
-
-};
-
-
-
-#endif //PROJECT_BENCHCOM_H
